requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.14.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cohere==5.18.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.118.0
fastavro==1.12.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.4
llama-index-cli==0.5.3
llama-index-core==0.14.4
llama-index-embeddings-cohere==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-openai==0.6.3
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-redis==0.6.1
llama-index-workflows==2.6.0
llama-parse==0.6.54
MarkupSafe==3.0.3
marshmallow==3.26.1
ml-dtypes==0.4.1
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
openai==1.109.1
pandas==2.2.3
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pypdf==6.1.1
pyreadline3==3.5.4
python-dotenv==1.1.1
python-ulid==3.1.0
pytz==2025.2
PyYAML==6.0.3
redis==5.3.1
redisvl==0.4.1
regex==2025.9.18
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
striprtf==0.0.26
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
twilio==9.8.3
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
aiofiles>=23.2.0